correct 1, 2, 5, 13, 14, 19, 20 point in backend and frontend project

+++ BackendTestTask/BackendTestTask/Controllers/CompaniesController.cs
@@ -3,6 +3,7 @@ using BackendTestTask.Models.Requests;
 using BackendTestTask.Services.CompanyService;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,9 +20,11 @@ namespace BackendTestTask.Controllers
     public class CompaniesController : ControllerBase
     {
         private readonly ICompanyService _companyService;
-        public CompaniesController(ICompanyService companyService)
+        private readonly ILogger<CompaniesController> _logger;
+        public CompaniesController(ICompanyService companyService, ILogger<CompaniesController> logger)
         {
             _companyService = companyService;
+            _logger = logger;
         }
 
         /// <summary>
@@ -31,15 +34,17 @@ namespace BackendTestTask.Controllers
         [HttpGet("get-companies")]
         public IActionResult GetCompanies()
         {
+            _logger.LogInformation("Starting get-companies request");
             try
             {
                 var companies = _companyService.GetCompanies();
 
                 if (companies == null)
                 {
+                    _logger.LogWarning("Failed to get-companies");
                     return BadRequest(new { message = "Не удалось получить компании" });
                 }
-
+                _logger.LogInformation("Returning a response to a request for get-companies");
                 return Ok(companies);
             }
             catch (Exception ex)
@@ -56,14 +61,24 @@ namespace BackendTestTask.Controllers
         [HttpPost("add-company")]
         public IActionResult AddCompany([FromBody] CompanyRequestModel model)
         {
+
+            if (!(model.Name.Length > 0 && model.Ticker.Length > 0))
+            {
+                return BadRequest();
+            }
+
+            _logger.LogInformation("Starting add-company request");
             try
             {
-                if (_companyService.AddCompany(model.Name, model.Ticker))
+                var company = _companyService.AddCompany(model.Name, model.Ticker);
+                if (company != null)
                 {
-                    return Ok(_companyService.GetCompanies());
+                    _logger.LogInformation("Returning a response to a request for add-company");
+                    return Ok(company);
                 }
                 else
                 {
+                    _logger.LogWarning("Failed to add-company");
                     return BadRequest("Не удалось добавить компанию");
                 }
             }
@@ -81,14 +96,23 @@ namespace BackendTestTask.Controllers
         [HttpPost("remove-company")]
         public IActionResult RemoveCompany([FromBody] CompanyRequestModel model)
         {
+
+            if (!(model.Name.Length > 0 && model.Ticker.Length > 0))
+            {
+                return BadRequest("Invalid input data");
+            }
+
+            _logger.LogInformation("Starting remove-company request");
             try
             {
                 if (_companyService.RemoveCompany(model.Id))
                 {
-                    return Ok(_companyService.GetCompanies());
+                    _logger.LogInformation("Returning a response to a request for remove-company");
+                    return Ok(model);
                 }
                 else
                 {
+                    _logger.LogWarning("Failed to remove-company");
                     return BadRequest("Не удалось удалить запись");
                 }
             }
@@ -106,14 +130,23 @@ namespace BackendTestTask.Controllers
         [HttpPost("change-company")]
         public IActionResult ChangeCompany([FromBody] CompanyRequestModel model)
         {
+
+            if (!(model.Name.Length > 0 && model.Ticker.Length > 0))
+            {
+                return BadRequest("Invalid input data");
+            }
+                
+            _logger.LogInformation("Starting change-company request");
             try
             {
                 if (_companyService.ChangeCompany(model.Id, model.Name, model.Ticker))
                 {
-                    return Ok(_companyService.GetCompanies());
+                    _logger.LogInformation("Returning a response to a request for change-company");
+                    return Ok(model);
                 }
                 else
                 {
+                    _logger.LogWarning("Failed to change-company");
                     return BadRequest("Не удалось изменить запись");
                 }
             }
